+++ cli/src/main/java/com/liferay/blade/cli/command/CreateArgs.java
@@ -72,6 +72,10 @@ public class CreateArgs extends BaseArgs {
 		return _hostBundleVersion;
 	}
 
+	public String getJSFramework() {
+		return _jsFramework;
+	}
+
 	public String getLiferayVersion() {
 		return _liferayVersion;
 	}
@@ -86,6 +90,10 @@ public class CreateArgs extends BaseArgs {
 
 	public String getPackageName() {
 		return _packageName;
+	}
+
+	public String getProduct() {
+		return _product;
 	}
 
 	public String getService() {
@@ -132,6 +140,10 @@ public class CreateArgs extends BaseArgs {
 		_hostBundleVersion = hostBundleVersion;
 	}
 
+	public void setJSFramework(String jsFramework) {
+		_jsFramework = jsFramework;
+	}
+
 	public void setLiferayVersion(String liferayVersion) {
 		_liferayVersion = liferayVersion;
 	}
@@ -146,6 +158,10 @@ public class CreateArgs extends BaseArgs {
 
 	public void setPackageName(String packageName) {
 		_packageName = packageName;
+	}
+
+	public void setProduct(String product) {
+		_product = product;
 	}
 
 	public void setService(String service) {
@@ -205,6 +221,12 @@ public class CreateArgs extends BaseArgs {
 	private String _hostBundleVersion;
 
 	@Parameter(
+		description = "Specify the javascript framework which will be used in the generated project. (metaljs)|(react)",
+		names = "--js-framework"
+	)
+	private String _jsFramework;
+
+	@Parameter(
 		description = "The version of Liferay to target when creating the project. Available options are 7.0, 7.1, 7.2, 7.3.",
 		names = {"-v", "--liferay-version"}
 	)
@@ -226,6 +248,9 @@ public class CreateArgs extends BaseArgs {
 
 	@Parameter(description = "The Java package to use when generating Java source.", names = {"-p", "--package-name"})
 	private String _packageName;
+
+	@Parameter(description = "The option for Liferay Platform product. (portal)|(dxp)", names = "--product")
+	private String _product = "portal";
 
 	@Parameter(
 		description = "If a new DS component needs to be created, provides the name of the service to be implemented.",