+++ cli/src/main/java/com/liferay/blade/cli/command/CreateCommand.java
@@ -50,7 +50,10 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
+import java.util.Optional;
 import java.util.Properties;
+
+import org.osgi.framework.Version;
 
 /**
  * @author Gregory Amerson
@@ -330,16 +333,20 @@ public class CreateCommand extends BaseCommand<CreateArgs> {
 		projectTemplatesArgs.setDependencyManagementEnabled(
 			(workspaceProvider != null) ? workspaceProvider.isDependencyManagementEnabled(dir) : false);
 
-		String liferayVersion = _getLiferayVersion(workspaceProvider, createArgs);
+		Optional<String> liferayVersion = _getLiferayVersion(workspaceProvider, createArgs);
 
-		if (BladeUtil.isEmpty(liferayVersion)) {
-			throw new IOException("LiferayVersion can not be Empty");
+		if (!liferayVersion.isPresent()) {
+			throw new IOException("Cannot determine Liferay Version. Please enter a valid value for Liferay Version.");
 		}
 
-		projectTemplatesArgs.setLiferayVersion(liferayVersion);
+		projectTemplatesArgs.setLiferayVersion(liferayVersion.get());
 
 		projectTemplatesArgs.setName(name);
 		projectTemplatesArgs.setPackageName(createArgs.getPackageName());
+
+		Optional<String> product = _getProduct(workspaceProvider, createArgs);
+
+		projectTemplatesArgs.setProduct(product.orElse(createArgs.getProduct()));
 
 		projectTemplatesArgs.setTemplate(template);
 
@@ -355,6 +362,7 @@ public class CreateCommand extends BaseCommand<CreateArgs> {
 		properties.put("setFrameworkDependencies", createArgs.getFrameworkDependencies());
 		properties.put("setHostBundleSymbolicName", createArgs.getHostBundleBSN());
 		properties.put("setHostBundleVersion", createArgs.getHostBundleVersion());
+		properties.put("setJSFramework", createArgs.getJSFramework());
 		properties.put("setOriginalModuleName", createArgs.getOriginalModuleName());
 		properties.put("setService", createArgs.getService());
 		properties.put("setViewType", createArgs.getViewType());
@@ -501,9 +509,15 @@ public class CreateCommand extends BaseCommand<CreateArgs> {
 		return _getDefaultDir(WorkspaceConstants.DEFAULT_WARS_DIR_PROPERTY, WorkspaceConstants.DEFAULT_WARS_DIR);
 	}
 
-	private String _getLiferayVersion(WorkspaceProvider workspaceProvider, CreateArgs createArgs) throws IOException {
+	private Optional<String> _getLiferayVersion(WorkspaceProvider workspaceProvider, CreateArgs createArgs)
+		throws IOException {
+
 		if (workspaceProvider == null) {
-			return createArgs.getLiferayVersion();
+			return Optional.ofNullable(
+				createArgs.getLiferayVersion()
+			).filter(
+				BladeUtil::isNotEmpty
+			);
 		}
 
 		File dir = createArgs.getDir();
@@ -512,13 +526,32 @@ public class CreateCommand extends BaseCommand<CreateArgs> {
 			dir = createArgs.getBase();
 		}
 
-		String liferayVersion = workspaceProvider.getLiferayVersion(dir);
+		String liferayVersion = createArgs.getLiferayVersion();
 
 		if (liferayVersion == null) {
-			return createArgs.getLiferayVersion();
+			liferayVersion = workspaceProvider.getLiferayVersion(dir);
 		}
 
-		return liferayVersion;
+		return _normalizeLiferayVersion(
+			Optional.ofNullable(
+				liferayVersion
+			).filter(
+				BladeUtil::isNotEmpty
+			));
+	}
+
+	private Optional<String> _getProduct(WorkspaceProvider workspaceProvider, CreateArgs createArgs) {
+		if (workspaceProvider == null) {
+			return Optional.empty();
+		}
+
+		File dir = createArgs.getDir();
+
+		if (dir == null) {
+			dir = createArgs.getBase();
+		}
+
+		return Optional.ofNullable(workspaceProvider.getProduct(dir));
 	}
 
 	private boolean _isExistingTemplate(String templateName) throws Exception {
@@ -531,6 +564,27 @@ public class CreateCommand extends BaseCommand<CreateArgs> {
 		BladeCLI bladeCLI = getBladeCLI();
 
 		return bladeCLI.isWorkspaceDir(dir);
+	}
+
+	private Optional<String> _normalizeLiferayVersion(Optional<String> liferayVersion) {
+		if (!liferayVersion.isPresent()) {
+			return Optional.empty();
+		}
+
+		Optional<String> formattedLiferayVersion = Optional.empty();
+
+		String versionValue = liferayVersion.get();
+
+		try {
+			Version version = Version.parseVersion(versionValue.replaceAll("-", "."));
+
+			formattedLiferayVersion = Optional.of(version.getMajor() + "." + version.getMinor());
+		}
+		catch (Exception exception) {
+			formattedLiferayVersion = Optional.of(versionValue.substring(0, 3));
+		}
+
+		return formattedLiferayVersion;
 	}
 
 }