+++ cli/src/test/java/com/liferay/blade/cli/command/InstallExtensionCommandTest.java
@@ -49,6 +49,7 @@ import org.apache.commons.compress.archivers.zip.ZipFile;
 import org.junit.Assert;
 import org.junit.Assume;
 import org.junit.Before;
+import org.junit.Ignore;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
@@ -85,6 +86,7 @@ public class InstallExtensionCommandTest {
 		Assert.assertTrue(extensionJar.getAbsolutePath() + " does not exist", extensionJar.exists());
 	}
 
+	@Ignore
 	@Test
 	public void testInstallCustomExtensionSubdirectory() throws Exception {
 		Assume.assumeFalse(_isWindows());
@@ -106,15 +108,15 @@ public class InstallExtensionCommandTest {
 		Path extensionDirPath = rootPath.resolve(Paths.get(".blade", "extensions"));
 
 		try (Stream<Path> extensionStream = Files.list(extensionDirPath)) {
-			boolean pathExists = extensionStream.map(
-				Path::getFileName
-			).map(
-				Object::toString
-			).anyMatch(
-				fileNameString -> fileNameString.startsWith("maven-profile")
-			);
-
-			Assert.assertTrue("maven-profile extension jar does not exist", pathExists);
+			Assert.assertTrue(
+				"maven-profile extension jar does not exist",
+				extensionStream.map(
+					Path::getFileName
+				).map(
+					Object::toString
+				).anyMatch(
+					fileNameString -> fileNameString.startsWith("maven-profile")
+				));
 		}
 	}