+++ extensions/project-templates-js-widget/src/main/java/com/liferay/project/templates/js/widget/internal/JSWidgetProjectTemplateCustomizer.java
@@ -61,8 +61,6 @@ public class JSWidgetProjectTemplateCustomizer implements ProjectTemplateCustomi
 
 		String workspaceLocation = ext.getWorkspaceLocation();
 
-		String modulesLocation = ext.getModulesLocation();
-
 		if (workspaceLocation != null) {
 			Path liferayLocationPath = Paths.get(workspaceLocation);
 
@@ -79,6 +77,8 @@ public class JSWidgetProjectTemplateCustomizer implements ProjectTemplateCustomi
 			config = _replace(config, "[$LIFERAY_DIR$]", liferayLocation);
 			config = _replace(config, "[$LIFERAY_PRESENT$]", "true");
 
+			String modulesLocation = ext.getModulesLocation();
+
 			File file = new File(modulesLocation, projectTemplatesArgs.getName());
 
 			modulesLocation = file.getAbsolutePath();
@@ -93,9 +93,7 @@ public class JSWidgetProjectTemplateCustomizer implements ProjectTemplateCustomi
 
 			workspacePath = workspacePath.normalize();
 
-			Path relativePath = workspacePath.relativize(modulesPath);
-
-			String relativePathString = relativePath.toString();
+			String relativePathString = String.valueOf(workspacePath.relativize(modulesPath));
 
 			if (OSDetector.isWindows()) {
 				relativePathString = relativePathString.replace("\\", "\\\\");